+++ models/comment.go
@@ -22,7 +22,7 @@ type Comment struct {
 	CommentID string `json:"comment_id,omitempty"`
 
 	// created at
-	CreatedAt int64 `json:"created_at,omitempty"`
+	CreatedAt float64 `json:"created_at,omitempty"`
 
 	// sort key
 	SortKey int64 `json:"sort_key,omitempty"`
